+++ e2e-tests/utils/php-src/TestDataBundles.php
@@ -93,7 +93,7 @@ final class TestDataBundles {
             propsData: ["filterPageType" => "Pages",
                 "filterLimit" => 0,
                 "filterOrder" => "desc",
-                "filterAdditional" => "[]"],
+                "filterAdditional" => "{}"],
             id: "@auto");
         return $out;
     }